requirements.txt
python-dotenv
requests
groq
psycopg2-binary
redis
fastapi
uvicorn
pydantic
